pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sigenergy2mqtt"
dynamic = ["version"]
dependencies = [
    "influxdb-client==1.49.0",
    "paho-mqtt==2.1.0",
    "psutil==7.2.1",
    "pymodbus==3.11.4",
    "requests==2.32.5",
    "ruamel.yaml==0.19.1",
    "scapy==2.7.0",
]
description = "Publish Modbus data from Sigenergy to MQTT, with optional Home Assistant Auto-Discovery and PVOutput updating"
authors = [
    { name = "seud0nym", email = "seud0nym@yahoo.com.au" }
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
readme = "README.md"
requires-python = ">=3.12"

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-xdist",
]

[tool.hatch.envs.default]
features = ["test"]
scripts.cov = "pytest --cov=sigenergy2mqtt --cov-report=xml --maxfail=0"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-n auto --dist loadscope"
markers = [
    "no_language_mock: disable global mock of i18n.get_default_language",
    "no_persistent_state_mock: disable global mock of Config.persistent_state_path",
]

[project.scripts]
sigenergy2mqtt = "sigenergy2mqtt.__main__:main"

[project.urls]
repository = "https://github.com/seud0nym/sigenergy2mqtt"
issues = "https://github.com/seud0nym/sigenergy2mqtt/issues"

[tool.hatch.version]
path = "sigenergy2mqtt/config/version.py"

[tool.hatch.build.targets.sdist]
include = ["sigenergy2mqtt/**"]

[tool.hatch.build.targets.wheel]
include = ["sigenergy2mqtt/**"]

[tool.pyright]
extraPaths = ["./sigenergy2mqtt"]
exclude = ["tests"]
ignore = ["tests"]
typeCheckingMode = "standard"
